Introduce a NixOS VM test for the curiosity server
This VM test mimicks the production setup and performs some curl-based
checks verifying that curiosity is able to serve the root page. We
also check that the Nginx setup is able to serve brotli-encoded static
content as well as performing some SSI operations.

The server setup is built on top of the production machine. We only
alter that configuration to inject a proper self-signed TLS
infrastracture with a /etc/hostname-based SNI setup.