Server: add setuid/setgid capabilities
We add tree new CLI flags to cty serve:

- --unix-socket-path: path pointing to the unix socket. Defaults to
  ./curiosity.
- --user-name: name of the user we want to run curiosity for.
- --group-name: name of the user we want to run curiosity for.

On a production setting, we often like to bind the unix sockets to
/run. You need to be root to do that. The idea is to start the
software as root, let it bind its socket in /run, then setuid/setgid
the process to its definitive user/group.

We decided not to change the state.json uid/gid. It means that we'll
have to "manually" set the appropriate permission. Likely via the
service pre-start procedure. It might be a bad idea we want to
revisit, it might be the right way to do it, only experience will
tell on that one.