Add a "public" version of the content.
The "public" version of the documentation pages uses a template where a
navigation bar is present, instead of an Nginx SSI comment (which thus
requires Nginx and the `cty serve` backend to fill in the navigation
bar).

That version can be built by using the `public` Nix attribute instead of
the `content` attribute.