Also use for Data.Business.
@noteed this can help us implement the ID show idea you were speaking
of in a more principled way.

If this is ok with you, I can proceed to add all instances for all IDs
that are prefixed.

The idea is that instead of having custom parsing functions, display
functions etc. for each ID (which is prefixed), we can use this
uniform insterface.