@noteed
This fixes the broken tests: for users.
I'm sorry but I had not assumed that the IDs were being stored in
newtypes with the prefixes.

This broke all tests and needed to be fixed.

This commit /should/ be able to fix them using the newtype machinery.

--

The earlier ID prefixing introduced some data redundancy.

For example:

```
UserId "USER-1"
```
is not very useful: it should not contain the prefix `USER-` inside
the newtype value, since the newtype guarantees that the Text is a
UserId. So we were storing redundant data.

Instead, the UserId, should be stored as `UserId "1"` and only
rendered as `USER-1` etc. when serialising (show, toJSON etc.) and
vice versa for parsing.

--

I'd love to have fixed this, but I ran out of time (and we had to
leave for the evening to friends.

My apologies for that. I'll try to finish this the coming week or the
next weekend.