CI: Introduce queued-build-hook
We use the queued-build-hook daemon to process the post build hooks.

This daemon executes asynchronously the post-build-hooks. It's a
significant performance improvement: the github action won't wait for
a path to be uploaded to the cache to build the next one but will
instead do that in the background.

At the end of the CI run, we wait for the queue daemon to upload all
the pending paths before stopping the CI run.