ci/build-curiosity.sh: fix error handling bug
We had a bug in the error handling path. The eval error status was
overriding the build error status in a pure bash style.

This was leading sucessful runs to return "null" (the eval error
status), failing the whole build.

We take advantage of this bugfix to properly display the derivation
pname in place of the attr name (which is always null in our setup: we
build a single derivation).