XXX Introdue an EmailTemplate type class.
The idea is to use a type class that can be instantiated by concrete
email template data types, instead of using a single sum type such as

    data EmailTemplate =
       SignupConfirmationEmail
     | InviteEmail
     | ...

The problem is that we want to serialize that to JSON and parse it back.
Serialization is easy. But parsing requires to know in a central place
how to map template names to concrete instances... So we're back to a
central place where all email templates are listed.

So we'll just put this commit aside for now.